refactor: 4계층 아키텍처 패키지 리팩토링
