+++ src/main/java/com/michelet/timeslotservice/application/service/TimeSlotService.java
@@ -1,11 +1,12 @@
-package com.michelet.timeslotservice.service;
+package com.michelet.timeslotservice.application.service;
 
 import com.michelet.common.exception.BusinessException;
 import com.michelet.timeslotservice.domain.TimeSlot;
-import com.michelet.timeslotservice.exception.TimeSlotErrorCode;
-import com.michelet.timeslotservice.repository.TimeSlotRepository;
-import com.michelet.timeslotservice.repository.entity.TimeSlotEntity;
-import com.michelet.timeslotservice.repository.mapper.TimeSlotMapper;
+import com.michelet.timeslotservice.domain.exception.TimeSlotErrorCode;
+import com.michelet.timeslotservice.infrastructure.config.persistence.TimeSlotRepository;
+import com.michelet.timeslotservice.infrastructure.config.persistence.entity.TimeSlotEntity;
+import com.michelet.timeslotservice.infrastructure.config.persistence.mapper.TimeSlotMapper;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
