+++ src/main/java/com/michelet/timeslotservice/infrastructure/config/persistence/entity/TimeSlotEntity.java
@@ -1,4 +1,4 @@
-package com.michelet.timeslotservice.repository.entity;
+package com.michelet.timeslotservice.infrastructure.config.persistence.entity;
 
 import com.michelet.common.entity.BaseEntity;
 import com.michelet.timeslotservice.domain.TimeSlotStatus;
